refactor : storage.js를 utils 폴더로 이동
